requirements.txt
# Core dependencies
python-dotenv>=0.19.0
pyyaml>=6.0
numpy>=1.21.0
pandas>=1.3.0

# Deep Learning
torch>=2.0.0
transformers>=4.30.0

# NLP
regex>=2022.10.31

# Data validation
pydantic>=1.10.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Logging and utils
tqdm>=4.64.0

# API (optional, for FastAPI demo)
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic[email]>=1.10.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
